// src/modul-2/fetch-2/interfaces/IHouse.ts
export interface IHouse {
  id: string
  name: string
  houseColours: string
  founder: string
  animal: string
  element: string
  ghost: string
  commonRoom: string
  heads: IHead[]
  traits: ITrait[]
}

export interface IHead {
  id: string
  firstName: string
  lastName: string
}

export interface ITrait {
  id: string
  name: string
}